% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getset.R
\name{AccessData4M}
\alias{AccessData4M}
\alias{identification}
\alias{identification<-}
\alias{observedLocations}
\alias{observedLocations<-}
\alias{interpolationParameters}
\alias{interpolationParameters<-}
\alias{interpolatedLocations}
\alias{interpolatedLocations<-}
\alias{movementData}
\alias{movementData<-}
\alias{groups}
\title{Get or set slots from an object of class \code{Data4M}.}
\usage{
identification(x)

identification(x) <- value

observedLocations(x)

observedLocations(x) <- value

interpolationParameters(x)

interpolationParameters(x) <- value

interpolatedLocations(x)

interpolatedLocations(x) <- value

movementData(x)

movementData(x) <- value

groups(x)
}
\arguments{
\item{x}{An object of class \code{Data4M}.}

\item{value}{A replacement value.}
}
\description{
Get or set slots from an object of class \code{Data4M}.
}
\seealso{
Other Access4M: \code{\link{AccessModel4M}},
  \code{\link{AccessSetModel4M}},
  \code{\link{AccessSetSim4M}},
  \code{\link{AccessSimulate4M}}
}
\concept{Access4M}
