% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getset.R
\docType{methods}
\name{AccessModel4M}
\alias{AccessModel4M}
\alias{SetModel4M}
\alias{parameterEstimates}
\alias{parameterEstimates<-}
\alias{aic}
\alias{aic<-}
\alias{residuals,Model4M-method}
\alias{residuals<-}
\alias{viterbiPath<-}
\alias{convergence}
\alias{convergence<-}
\alias{tmbEnvironment}
\alias{tmbEnvironment<-}
\title{Get or set slots of an object of class \code{Model4M}.}
\usage{
SetModel4M(x)

parameterEstimates(x)

parameterEstimates(x) <- value

aic(x)

aic(x) <- value

\S4method{residuals}{Model4M}(object)

residuals(x) <- value

viterbiPath(x) <- value

convergence(x)

convergence(x) <- value

tmbEnvironment(x)

tmbEnvironment(x) <- value
}
\arguments{
\item{value}{A replacement value.}

\item{x/object}{An object of class \code{Model4M}.}
}
\description{
Get or set slots of an object of class \code{Model4M}.
}
\seealso{
Other Access4M: \code{\link{AccessData4M}},
  \code{\link{AccessSetModel4M}},
  \code{\link{AccessSetSim4M}},
  \code{\link{AccessSimulate4M}}
}
\concept{Access4M}
