% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getset.R
\name{AccessSetModel4M}
\alias{AccessSetModel4M}
\alias{nStates}
\alias{nStates<-}
\alias{useHMM}
\alias{useHMM<-}
\alias{distribution}
\alias{distribution<-}
\alias{zeroInflation}
\alias{zeroInflation<-}
\alias{startingValues}
\alias{startingValues<-}
\alias{parameterMapping}
\alias{parameterMapping<-}
\title{Get or set slots of an object of class \code{SetModel4M}.}
\usage{
nStates(x)

nStates(x) <- value

useHMM(x)

useHMM(x) <- value

distribution(x)

distribution(x) <- value

zeroInflation(x)

zeroInflation(x) <- value

startingValues(x)

startingValues(x) <- value

parameterMapping(x)

parameterMapping(x) <- value
}
\arguments{
\item{x}{An object of class \code{SetModel4M}.}

\item{value}{A replacement value.}
}
\description{
Get or set slots of an object of class \code{SetModel4M}.
}
\seealso{
Other Access4M: \code{\link{AccessData4M}},
  \code{\link{AccessModel4M}},
  \code{\link{AccessSimulate4M}}
}
