% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getset.R
\docType{methods}
\name{AccessSetSim4M}
\alias{AccessSetSim4M}
\alias{simLength}
\alias{simLength<-}
\alias{nStates,SetSim4M-method}
\alias{nStates<-,SetSim4M-method}
\alias{useHMM,SetSim4M-method}
\alias{useHMM<-,SetSim4M-method}
\alias{distribution,SetSim4M-method}
\alias{parameters}
\alias{parameters<-}
\title{Get or set slots of an object of class \code{SetSim4M}.}
\usage{
simLength(x)

simLength(x) <- value

\S4method{nStates}{SetSim4M}(x)

\S4method{nStates}{SetSim4M}(x) <- value

\S4method{useHMM}{SetSim4M}(x)

\S4method{useHMM}{SetSim4M}(x) <- value

\S4method{distribution}{SetSim4M}(x)

parameters(x)

parameters(x) <- value
}
\arguments{
\item{x}{An object of class \code{SetSim4M}.}

\item{value}{A replacement value.}
}
\description{
Get or set slots of an object of class \code{SetSim4M}.
}
\seealso{
Other Access4M: \code{\link{AccessData4M}},
  \code{\link{AccessModel4M}},
  \code{\link{AccessSetModel4M}},
  \code{\link{AccessSimulate4M}}
}
\concept{Access4M}
