% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getset.R
\docType{methods}
\name{AccessSimulate4M}
\alias{AccessSimulate4M}
\alias{simulatedLocations}
\alias{simulatedLocations<-}
\alias{simulatedData}
\alias{simulatedData<-}
\alias{simulatedViterbiPath}
\alias{simulatedViterbiPath<-}
\alias{refitParameters}
\alias{refitParameters<-}
\alias{refitResiduals}
\alias{refitResiduals<-}
\alias{refitViterbiPath}
\alias{refitViterbiPath<-}
\alias{refitConvergence}
\alias{refitConvergence<-}
\alias{refitEnvironment}
\alias{refitEnvironment<-}
\alias{length,Simulate4M-method}
\title{Get or set slots of an object of class \code{Simulate4M}.}
\usage{
simulatedLocations(x)

simulatedLocations(x) <- value

simulatedData(x)

simulatedData(x) <- value

simulatedViterbiPath(x)

simulatedViterbiPath(x) <- value

refitParameters(x)

refitParameters(x) <- value

refitResiduals(x)

refitResiduals(x) <- value

refitViterbiPath(x)

refitViterbiPath(x) <- value

refitConvergence(x)

refitConvergence(x) <- value

refitEnvironment(x)

refitEnvironment(x) <- value

\S4method{length}{Simulate4M}(x)
}
\arguments{
\item{x}{An object of class \code{Simulate4M}.}

\item{value}{A replacement value.}
}
\description{
Get or set slots of an object of class \code{Simulate4M}.
}
\seealso{
Other Access4M: \code{\link{AccessData4M}},
  \code{\link{AccessModel4M}},
  \code{\link{AccessSetModel4M}},
  \code{\link{AccessSetSim4M}}
}
\concept{Access4M}
