% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\docType{class}
\name{Data4M-class}
\alias{Data4M-class}
\title{An S4 class to hold and pre-process movement data.}
\description{
An S4 class to hold and pre-process movement data.
}
\section{Slots}{

\describe{
\item{\code{Identification}}{A string, list, number, etc. used for identification.}

\item{\code{Observed.Locations}}{A data frame with columns \code{Date}, \code{Lon}, and \code{Lat}.}

\item{\code{Interpolation.Parameters}}{A named vector with positive entries \code{Time.Step} and \code{Group.Cutoff}.}

\item{\code{Interpolated.Locations}}{A data frame with columns \code{Date}, \code{Lon}, \code{Lat}, and \code{Group}.}

\item{\code{Movement.Data}}{A named list holding the pre-processed data with the following entries:
\describe{
  \item{Movement.Data}{A data frame with columns \code{Deflection.Angle}, \code{Step.Length}, and \code{Group}.}
  \item{Step.Length.Starting.Values}{A numeric vector giving the first step length for each group.}
  \item{Step.Length.Mean}{The mean of the unstandardized step lengths.}
}}
}}

\seealso{
Other 4M-classes: \code{\link{Model4M-class}},
  \code{\link{SetModel4M-class}},
  \code{\link{SetSim4M-class}},
  \code{\link{Simulate4M-class}}
}
\concept{4M-classes}
