% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getset.R
\docType{methods}
\name{data4M}
\alias{data4M}
\alias{initialize,Data4M-method}
\alias{data4M,data.frame-method}
\alias{data4M,Model4M-method}
\title{Create or extract a \code{Data4M} object.}
\usage{
data4M(x, ...)

\S4method{initialize}{Data4M}(.Object, Identification = list(),
  Data = data.frame(Date = .POSIXct(numeric(0)), Lon = numeric(0), Lat =
  numeric(0)))

\S4method{data4M}{data.frame}(x, Identification = list())

\S4method{data4M}{Model4M}(x)
}
\arguments{
\item{x}{An object to coerce to class \code{Data4M}. If a data.frame, it must contain at least the columns \code{Date}, \code{Longitude}, and \code{Latitude}. These column names can be shortened e.g., \code{Lat}.}

\item{Identification}{An optional argument of any type to help the user identify the data set.}
}
\description{
Takes a data.frame of time-stamped locations and store it in an object of class \code{Data4M} for further processing. Otherwise takes an object and coerces it to class Data4M.
}
\details{
The \code{initialize} function is not meant to be used by the user, use \code{data4M} instead.
}
\seealso{
Other Construct4M: \code{\link{model4M}},
  \code{\link{setModel4M}}, \code{\link{setSim4M}},
  \code{\link{simulate4M}}
}
\concept{Construct4M}
