% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genmet.R
\docType{methods}
\name{difftime4M}
\alias{difftime4M}
\title{Compute time differences in the observed locations of an object of class \code{Data4M}.}
\usage{
\S4method{difftime}{Data4M}(time1, units = "hours")
}
\arguments{
\item{time1}{An object of class \code{Data4M}.}

\item{units}{Units in which the result is desired. See \code{\link{difftime}}.}
}
\value{
Returns a numeric vector of time differences between observed locations.
}
\description{
Compute time differences in the observed locations of an object of class \code{Data4M}.
}
\examples{
sealdata<- data4M(greyseal)
summary(difftime(sealdata))

}
