% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genmet.R
\docType{methods}
\name{findStep}
\alias{findStep}
\alias{findStep,Data4M-method}
\title{Compute diagnostic statistics for a grid of time steps and group cutoffs.}
\usage{
findStep(object, ...)

\S4method{findStep}{Data4M}(object,
  step.grid = seq(summary(difftime(object))[[3]],
  summary(difftime(object))[[5]], length.out = 10), cutoff.grid = seq(1,
  2, length.out = 11), units = "hrs")
}
\arguments{
\item{object}{An object of or inheriting from class \code{Data4M}.}

\item{step.grid}{A numeric vector giving the values of Time.Step to consider.}

\item{cutoff.grid}{A numeric vector giving the multiples of Time.Step to use as the Group.Cutoff (all entries should be between 1 and 2).}
}
\value{
A list giving the values of n_prop and n_adj, along with the values which maximize each statistic.
}
\description{
Compute diagnostic statistics for a grid of time steps and group cutoffs.
}
\examples{
sealdata<- data4M(greyseal)
findStep(sealdata,cutoff.grid = seq(1,2,length.out = 5)

}
