% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{greyseal}
\alias{greyseal}
\title{A movement track.}
\format{A data frame with 3,158 rows and 3 variables:
\describe{
  \item{date}{Date, given in \%Y-\%m-\%d \%H:\%M:\%S.}
  \item{lon}{Longitude, in decimal degrees.}
  \item{lat}{Latitude, in decimal degrees.}
}}
\source{
\url{https://onlinelibrary.wiley.com/doi/full/10.1002/ece3.2795}
}
\usage{
greyseal
}
\description{
A dataset containing 3,158 time-stamped locations of a male grey seal in the Atlantic ocean near the Scotian shelf.
}
\keyword{datasets}
