% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genmet.R
\docType{methods}
\name{interpolate}
\alias{interpolate}
\alias{interpolate,Data4M-method}
\title{Linearly interpolate and group locations to a regular time step.}
\usage{
interpolate(object, ...)

\S4method{interpolate}{Data4M}(object, Time.Step, Group.Cutoff = 2 *
  Time.Step, units = "hrs")
}
\arguments{
\item{object}{An object of or inheriting from class \code{Data4M}.}

\item{Time.Step}{The time interval used in the regular time grid.}

\item{Group.Cutoff}{Observations which are spaced further apart than this will be separated into different group.}

\item{units}{Can be one of \code{days}, \code{hrs}, \code{mins}, or \code{secs}.}
}
\value{
Returns a copy of \code{object} with the Interpolation.Parameters, Interpolated.Locations, and Movement.Data slots filled in.
}
\description{
Provide a time interval and cutoff level to linearly interpolate a movement track.
}
\examples{
sealdata<- data4M(greyseal)
sealdata<- interpolate(sealdata,Time.Step = 66, units = "mins")

}
