% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\docType{class}
\name{Model4M-class}
\alias{Model4M-class}
\title{An S4 class for fitted models from \linkS4class{Data4M} and \linkS4class{SetModel4M}.}
\description{
An S4 class for fitted models from \linkS4class{Data4M} and \linkS4class{SetModel4M}.
}
\section{Slots}{

\describe{
\item{\code{Parameter.Estimates}}{A named list holding parameter estimates with the following entries:
\describe{
  \item{Deflection.Angle.Parameters}{A data.frame with columns \code{Center} and \code{Concentration}.}
  \item{Step.Length.Parameters}{A data.frame with columns \code{Intercept}, \code{Autocorrelation}, and \code{Standard.Deviation}.}
  \item{Transition.Probabilities}{A square matrix of transition probabilities.}
  \item{Stationary.Distribution}{A numeric vector giving the stationary distribution of \code{Transition.Probabilties}.}
  \item{Zero.Inflation}{A named list containing zero-inflation parameters.}
}}

\item{\code{AIC}}{A numeric value giving the AIC of the fitted model.}

\item{\code{Residuals}}{A data.frame with columns \code{Deflection.Angle} and \code{Step.Length} giving probability-scale forecast residuals.}

\item{\code{Viterbi.Path}}{A integer vector giving the Viterbi decoded state estimates.}

\item{\code{Convergence}}{A character vector giving the convergence message.}

\item{\code{TmbEnvironment}}{The environment code for the TMB \code{MakeADFun} object.}
}}

\section{Extends}{

\code{\link{Data4M-class}}, \code{\link{SetModel4M-class}}.
}

\seealso{
Other 4M-classes: \code{\link{Data4M-class}},
  \code{\link{SetModel4M-class}},
  \code{\link{Simulate4M-class}}
}
