% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getset.R
\docType{methods}
\name{model4M}
\alias{model4M}
\alias{initialize,Model4M-method}
\alias{model4M,Data4M-method}
\alias{model4M,Simulate4M-method}
\title{Create or extract a \code{Model4M} object.}
\usage{
model4M(x, ...)

\S4method{initialize}{Model4M}(.Object, Data4M = new("Data4M"),
  SetModel4M = new("SetModel4M"))

\S4method{model4M}{Data4M}(x, Set.Model)

\S4method{model4M}{Simulate4M}(x)
}
\arguments{
\item{x}{An object to coerce to class \code{Model4M}, possibly of class \code{Data4M}.}

\item{SetModel4M}{An object of class \code{SetModel4M}.}
}
\description{
Create or extract a \code{Model4M} object.
}
\details{
Consider using \code{fit} to create a new \code{Model4M} object.
}
\seealso{
Other Construct4M: \code{\link{data4M}},
  \code{\link{setModel4M}}, \code{\link{setSim4M}},
  \code{\link{simulate4M}}
}
\concept{Construct4M}
