% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/showplot.R
\docType{methods}
\name{plot4M}
\alias{plot4M}
\alias{plot,Data4M,character-method}
\alias{plot,Model4M,character-method}
\title{Plot objects of class \code{Data4M} or \code{Model4M}.}
\usage{
\S4method{plot}{Data4M,character}(x, y, ask = interactive(),
  Viterbi.Path = NA, N.States = 0, h = 0.25, ...)

\S4method{plot}{Model4M,character}(x, y, h = 0.25, ask = interactive(),
  palette = "Set1", ...)
}
\arguments{
\item{x}{An object of class \code{Data4M} or \code{Model4M}.}

\item{y}{A character vector choosing which plots to draw. Can be any combination of \code{all}, \code{satellite}, \code{locations}, \code{data}, and \code{residuals}. Defaults to \code{all}.}

\item{ask}{Wait for input to switch to next plot?}

\item{Viterbi.Path}{State path used to color locations. Not typically entered by the user.}

\item{N.States}{The number of states used in the model. Not typically entered by the user.}

\item{h}{Vector of bandwidth for residual density plots. See \code{\link[MASS]{kde2d}}.}

\item{...}{Optional arguments used to control internal plotting functions, such as \code{pch} or \code{main}. Other useful arguments include...
\describe{
  \item{lag}{For data lag-plot, which time lag should be used? Defaults to 1.}
  \item{log}{For data lag-plot, should the log of step lengths be used? Defaults to FALSE.}
  \item{N.Grid.Lines}{How many grid lines should be used for each axis in the mesh for the data plots? Defaults to 100.}
  \item{N.Greys}{What greyscale depth to use for density plots of the data. Defaults to 25.}
}}

\item{palette}{If using package \code{RColorBrewer} and not supplying \code{col}, the palette to use for colors.}
}
\description{
Plot objects of class \code{Data4M} or \code{Model4M}.
}
\examples{
sealdata<- data4M(greyseal)
sealdata<- interpolate(sealdata,Time.Step = 1)
plot(sealdata,ask = F)
plot(sealdata,"data",log = T,lag = 2,N.Grid.Lines = 150,N.Greys = 30)
plot(sealdata,"locations",pch = 11,lty = 4,col = "red")

seal4M3<- fit(sealdata,N.States = 3)
plot(seal4M3,ask = F)
plot(seal4M3,"residuals")

library(RColorBrewer)
plot(seal4M3,"locations",palette = "Dark2")

}
