% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genmet.R
\docType{methods}
\name{preACF}
\alias{preACF}
\alias{preACF,Data4M-method}
\title{Estimate the autocorrelation function before interpolating locations.}
\usage{
preACF(object, ...)

\S4method{preACF}{Data4M}(object, Time.Step, units = "hrs",
  lag.max = 30, kernel = "gaussian", kernel.width = "default")
}
\arguments{
\item{object}{An object of or inheriting from class \code{Data4M}.}

\item{Time.Step}{The time interval used for the regular time grid.}

\item{units}{Can be one of \code{days}, \code{hrs}, \code{mins}, or \code{secs}.}

\item{lag.max}{The maximum lag to compute.}

\item{kernel.width}{The bandwidth of the kernel used in weighting pairs of observations. Note that the time differences are standardized to have mean 1. The default setting depends on the kernel used.}

\item{kernel.type}{Which kernel to use? Can be one of "rectangle", "sinc", or "gaussian". Can also supply a user-defined function that takes the arguments d, kernel.width, and n.}
}
\value{
An object of class \code{acf}, which is a list containing the following elements:
 \itemize{
     \item lag: A vector containing the lags at which the acf is estimated.
     \item acf: An vector with the same length as 'lag' containing the estimated acf.
     \item type: Always set to  "correlation".
     \item n.used A vector giving the sum of kernel values for each lag.
     \item series The name of the series 'object'.
     \item snames Always set to NULL.
 }
}
\description{
Estimate the acf for step length before interpolating the data.
}
