% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\docType{class}
\name{SetModel4M-class}
\alias{SetModel4M-class}
\title{An S4 class to hold model definition and fitting parameters.}
\description{
An S4 class to hold model definition and fitting parameters.
}
\section{Slots}{

\describe{
\item{\code{N.States}}{The number of states \code{n} used in the model.}

\item{\code{Use.HMM}}{If \code{TRUE} fixes autocorrelation to zero.}

\item{\code{Distribution}}{Distribution for step length. Can be one of \code{gamma} or \code{log-normal}.}

\item{\code{Zero.Inflation}}{A named logical vector with entries \code{Step.Length} and \code{Deflection.Angle}. A \code{TRUE} entry enables zero inflation.}

\item{\code{Starting.Values}}{A named list giving the starting values of the working parameters with the following entries:
\describe{
  \item{Tpm.Working.Pars}{An \code{n} by \code{n} matrix.}
  \item{Theta.Working.Pars}{An \code{n} by 2 matrix. The first column is untransformed center, the second is the logit of concentration.}
  \item{Step.Working.Pars}{An \code{n} by 3 matrix. The first column is the log of the intercept, the second is the log of the autocorrelation, the third is the log of the standard deviation.}
  \item{Logit.Step.Zero.Probs}{A length \code{n} vector.}
  \item{Logit.Angle.Zero.Probs}{A length \code{n} vector.}
  \item{Angle.Zero.Working.Pars}{An \code{n} by 2 matrix.}
}}

\item{\code{Parameter.Mapping}}{A named list with the same structure as \code{Starting.Values}. Controls TMB parameter mapping. See details of \code{\link[TMB]{MakeADFun}}.}
}}

\seealso{
Other 4M-classes: \code{\link{Data4M-class}},
  \code{\link{Model4M-class}},
  \code{\link{SetSim4M-class}},
  \code{\link{Simulate4M-class}}
}
\concept{4M-classes}
