% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\docType{class}
\name{SetSim4M-class}
\alias{SetSim4M-class}
\title{An S4 class to hold model definition parameters for simulation.}
\description{
An S4 class to hold model definition parameters for simulation.
}
\section{Slots}{

\describe{
\item{\code{Sim.Length}}{A positive integer giving the number of observations to simulate.}

\item{\code{N.States}}{The number of states \code{n} used in the model.}

\item{\code{Use.HMM}}{If \code{TRUE} fixes autocorrelation to zero.}

\item{\code{Distribution}}{Distribution for step length. Can be one of \code{gamma} or \code{log-normal}.}

\item{\code{Parameters}}{A named list giving the parameter values to simulate from:
\describe{
  \item{Deflection.Angle.Parameters}{A data.frame with columns \code{Center} and \code{Concentration}.}
  \item{Step.Length.Parameters}{A data.frame with columns \code{Intercept}, \code{Autocorrelation}, and \code{Standard.Deviation}.}
  \item{Transition.Probabilities}{A square matrix of transition probabilities.}
}}
}}

\seealso{
Other 4M-classes: \code{\link{Data4M-class}},
  \code{\link{Model4M-class}},
  \code{\link{SetModel4M-class}},
  \code{\link{Simulate4M-class}}
}
\concept{4M-classes}
