% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getset.R
\docType{methods}
\name{setSim4M}
\alias{setSim4M}
\alias{initialize,SetSim4M-method}
\title{Create or extract a \code{SetSim4M} object.}
\usage{
setSim4M(x, ...)

\S4method{initialize}{SetSim4M}(.Object, Sim.Length = 100,
  N.States = 2, Use.HMM = FALSE, Distribution = "gamma",
  Parameters = list())
}
\arguments{
\item{x}{An object to coerce to class \code{SetSim4M}. Missing in calls to create a new object.}
}
\description{
Creates or extracts an object used to hold a model definition.
}
\details{
The \code{initialize} function is not meant to be used by the user, use \code{setSim4M} instead.
}
\seealso{
Other Construct4M: \code{\link{data4M}},
  \code{\link{model4M}}, \code{\link{setModel4M}},
  \code{\link{simulate4M}}
}
\concept{Construct4M}
