% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genmet.R
\docType{methods}
\name{simCheck}
\alias{simCheck}
\alias{simCheck,Simulate4M-method}
\title{Check simulations for undetected non-convergence.}
\usage{
simCheck(x, ...)

\S4method{simCheck}{Simulate4M}(x, Stationary.Distribution.Check = 0.05,
  Angle.Concentration.Check = 0.05, Tpm.Tolerance = 4)
}
\arguments{
\item{x}{An object of class Simulate4M}

\item{Stationary.Distribution.Check}{Any simulation which has a stationary distribution having any element less than this value will be flagged.}

\item{Angle.Concentration.Check}{Any simulation which has a deflection angle concentration less than this value will be flagged.}

\item{Tpm.Tolerance}{Any simulation which has a transition probability matrix which contains a row with at least two entries which
are equal up to this many decimal places will be flagged.}
}
\value{
Returns an index vector which flags simulations which violate conditions.
}
\description{
Check simulations for undetected non-convergence.
}
