% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genmet.R
\docType{methods}
\name{simulate.4M}
\alias{simulate.4M}
\alias{simulate,Model4M-method}
\title{Simulate data from an object of class \code{Model4M}.}
\usage{
\S4method{simulate}{Model4M}(object, nsim = 1, seed = NULL,
  refit = T, keep.locations = T, keep.data = T, keep.residuals = T,
  Use.HMM = useHMM(object), Set.Model.4M = SetModel4M(object), ...)
}
\arguments{
\item{object}{An object of class \code{Model4M}.}

\item{nsim}{How many simulations should there be?}

\item{seed}{Sets the RNG seed. If NULL, does not alter the current RNG settings.}

\item{refit}{Should a model be refit to each set of simulated data?}

\item{keep.locations}{Should the simulated locations be kept?}

\item{keep.data}{Should the simulated movement data be kept?}

\item{keep.residuals}{If \code{refit = T}, should the residuals be kept?}

\item{Use.HMM}{If \code{refit = T}, should an HMM be fit? Overrides the value stored in \code{model} and \code{Set.Model.4M}.}

\item{Set.Model.4M}{If \code{refit = T}, which SetModel4M object should be used to refit the data? Use.HMM is controlled separately.}

\item{...}{Used to pass arguments to internal functions.}
}
\value{
Returns an object of class \code{Simulate4M} containing simulations and/or results of refitting the simulations.
}
\description{
Simulate data from an object of class \code{Model4M}.
}
\examples{
sealData<- data4M(greyseal)
sealData<- interpolate(sealData,Time.Step = 1)
seal4M2<- fit(sealData)
sim4M2<- simulate(seal4M2,nsim = 5)

}
