% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\docType{class}
\name{Simulate4M-class}
\alias{Simulate4M-class}
\title{An S4 class for holding simulations from a \code{\link{Model4M}} class.}
\description{
An S4 class for holding simulations from a \code{\link{Model4M}} class.
}
\section{Slots}{

\describe{
\item{\code{Simulated.Locations}}{A 3d array holding the simulated locations with columns \code{Date}, \code{Lon}, \code{Lat}, and \code{Group}, and depth corresponding to individual simulations.}

\item{\code{Simulated.Data}}{A 3d array holding the simulated movement data with columns \code{Deflection.Angle}, \code{Step.Length}, and \code{Group}, and depth corresponding to individual simulations.}

\item{\code{Simulated.Viterbi.Path}}{A matrix holding the simulated state path with rows corresponding to individual simulations.}

\item{\code{Refit.Parameters}}{A named list holding refit parameter estimates with the following entries:
\describe{
  \item{Deflection.Angle.Parameters}{A 3d array with rows corresponding to states, columns corresponding to parameters, and depth corresponding to individual simulations.}
  \item{Step.Length.Parameters}{A 3d array with rows corresponding to states, columns corresponding to parameters, and depth corresponding to individual simulations.}
  \item{Transition.Probabilities}{A 3d array with depth corresponding to individual simulations.}
  \item{Stationary.Distribution}{A matrix with rows corresponding to individual simulations and column corresponding to state.}
  \item{Zero.Inflation}{A named list containing zero-inflation parameters.}
}}

\item{\code{Refit.Residuals}}{A 3d array with columns \code{Deflection.Angle} and \code{Step.Length}, and depth corresponding to individual simulations.}

\item{\code{Refit.Viterbi.Path}}{A matrix holding the refit Viterbi-estimated state path with rows corresponding to individual simulations.}

\item{\code{Refit.Convergence}}{A character vector holding the convergence messages from each simulation.}

\item{\code{Refit.Environment}}{A list holding the TMB environments for each simulation.}
}}

\section{Extends}{

\code{\link{Model4M-class}}.
}

\seealso{
Other 4M-classes: \code{\link{Data4M-class}},
  \code{\link{Model4M-class}},
  \code{\link{SetModel4M-class}}
}
