% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genmet.R
\name{simulationError}
\alias{simulationError}
\alias{rmse}
\alias{mae}
\alias{bias}
\alias{stateError}
\title{Compute simulation parameter errors.}
\usage{
rmse(x)

mae(x)

bias(x)

stateError(x)
}
\arguments{
\item{x}{An object of class \code{Simulate4M}.}
}
\value{
Functions \code{rmse} and \code{mae} return a list of parameter errors. Function \code{stateError} returns a numeric vector containing the percentage of state estimates that are not correct.
}
\description{
Compute simulation parameter errors.
}
\examples{
sealData<- data4M(greyseal)
sealData<- interpolate(sealData,Time.Step = 1)
seal4M2<- fit(sealData)
sim4M2<- simulate(seal4M2,nsim = 5)
rmse(sim4M2)
stateError(sim4M2)

}
