% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getset.R
\docType{methods}
\name{SubsetCombineSimulate4M}
\alias{SubsetCombineSimulate4M}
\alias{[,Simulate4M,numeric,ANY,missing-method}
\alias{[[,Simulate4M,numeric,missing-method}
\alias{c,Simulate4M-method}
\title{Subset or combine objects of class \code{Simulate4M}.}
\usage{
\S4method{[}{Simulate4M,numeric,ANY,missing}(x, i, j, drop)

\S4method{[[}{Simulate4M,numeric,missing}(x, i, j)

\S4method{c}{Simulate4M}(x, ...)
}
\arguments{
\item{x}{An object of class \code{Simulate4M}.}

\item{i}{Simulations to extract.}

\item{...}{Additional objects of class Simulate4M with the same structure as \code{x}.}
}
\description{
Subset or combine objects of class \code{Simulate4M}.
}
